seperate types from useDashBoard, use assets from chain-registry
