+++ src/pages/index.vue
@@ -1,6 +1,7 @@
 <script lang="ts" setup>
 import { Icon } from '@iconify/vue';
-import { useDashboard, LoadingStatus, type ChainConfig } from '@/stores/useDashboard';
+import { useDashboard, LoadingStatus } from '@/stores';
+import type { ChainConfig } from '@/types/chaindata';
 import ChainSummary from '@/components/ChainSummary.vue';
 import AdBanner from '@/components/ad/AdBanner.vue';
